# api_specification.yaml
openapi: 3.0.3
info:
  title: Property Upkeep Records API
  description: |
    REST API for Property Upkeep Records System - A comprehensive solution for tracking property maintenance, repairs, and improvements.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Multi-tenancy
    All data is isolated by organization. Users can only access data for organizations they belong to.
    
    ## Rate Limiting
    API requests are rate limited to 100 requests per minute per user.
    
    ## Pagination
    List endpoints support cursor-based pagination with `limit` and `cursor` parameters.
  version: 1.0.0
  contact:
    name: Property Upkeep Records API Support
    email: support@propertyupkeep.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.propertyupkeep.com/v1
    description: Production server
  - url: https://staging-api.propertyupkeep.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Authenticate user with email and password, returning JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "manager@demo.com"
                password:
                  type: string
                  format: password
                  example: "SecurePassword123!"
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Use refresh token to get a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate current refresh token
      responses:
        '204':
          description: Successfully logged out
        '401':
          $ref: '#/components/responses/Unauthorized'

  # User endpoints
...
  /organizations:
    get:
      tags:
        - Organizations
      summary: List organizations
      description: List all organizations the current user has access to
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /properties:
    get:
      tags:
        - Properties
      summary: List properties
      description: List all properties for the current user's organization
      parameters:
        - name: search
          in: query
          description: Search properties by name or address
          schema:
            type: string
            example: "Oak Street"
        - name: property_type
          in: query
          description: Filter by property type
          schema:
            $ref: '#/components/schemas/PropertyType'
        - name: is_active
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: List of properties
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Property'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Properties
      summary: Create property
      description: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreate'
      responses:
        '201':
          description: Property created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /reports/work-summary:
    get:
      tags:
        - Reports
      summary: Work summary report
      description: Generate a summary report of work logs with various filters and groupings
      parameters:
        - name: property_id
          in: query
          description: Filter by specific property
          schema:
            type: string
            format: uuid
        - name: from_date
          in: query
          description: Filter work logs from this date
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - name: to_date
          in: query
          description: Filter work logs to this date
          schema:
            type: string
            format: date
            example: "2024-12-31"
        - name: category
          in: query
          description: Filter by work log category
          schema:
            $ref: '#/components/schemas/WorkCategory'
        - name: status
          in: query
          description: Filter by work log status
          schema:
            $ref: '#/components/schemas/WorkStatus'
        - name: group_by
          in: query
          description: Group results by field
          schema:
            type: string
            enum: [property, category, status, month]
            default: property
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Work summary report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkSummaryReport'
            text/csv:
              schema:
                type: string
                example: |
                  Property Name,Category,Status,Count,Total Cost,Avg Cost
                  Oak Street Apartments,MAINTENANCE,COMPLETED,15,2500.00,166.67
                  Oak Street Apartments,REPAIR,COMPLETED,8,1200.00,150.00
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Number of items to return (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      description: Pagination cursor for next page
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "bad_request"
            message: "Invalid request format"
            details:
              field: "email"
              code: "invalid_format"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Authentication required"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "forbidden"
            message: "Insufficient permissions to access this resource"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
...
    PropertyType:
      type: string
      enum: [RESIDENTIAL, COMMERCIAL, MIXED_USE]
...
    WorkCategory:
      type: string
      enum: [MAINTENANCE, REPAIR, IMPROVEMENT, INSPECTION]
...
    WorkStatus:
      type: string
      enum: [PLANNED, IN_PROGRESS, COMPLETED, CANCELLED]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        phone:
          type: string
          maxLength: 20
          nullable: true
        is_active:
          type: boolean
        email_verified:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, email, first_name, last_name, is_active, email_verified, created_at, updated_at]
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        type:
          type: string
          maxLength: 50
        settings:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, type, settings, created_at, updated_at]
    Property:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        address:
          $ref: '#/components/schemas/Address'
        property_type:
          $ref: '#/components/schemas/PropertyType'
        metadata:
          type: object
          additionalProperties: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
      required: [id, organization_id, name, address, property_type, is_active, created_at, updated_at]
    PropertyCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          example: "Oak Street Apartments"
        address:
          $ref: '#/components/schemas/Address'
        property_type:
          $ref: '#/components/schemas/PropertyType'
        metadata:
          type: object
          additionalProperties: true
          example:
            bedrooms: 24
            units: 12
            year_built: 1995
      required: [name, address, property_type]
...
    Address:
      type: object
      properties:
        street:
          type: string
          maxLength: 255
          example: "123 Oak Street"
        street2:
          type: string
          maxLength: 255
          nullable: true
          example: "Apt 4B"
        city:
          type: string
          maxLength: 100
          example: "London"
        state:
          type: string
          maxLength: 100
          nullable: true
          example: "England"
        postcode:
          type: string
          maxLength: 20
          example: "SW1A 1AA"
        country:
          type: string
          maxLength: 100
          example: "United Kingdom"
        latitude:
          type: number
          format: float
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
      required: [street, city, postcode, country]
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_in:
          type: integer
          example: 900
          description: Access token expiration time in seconds
        token_type:
          type: string
          example: "Bearer"
        user:
          $ref: '#/components/schemas/User'
      required: [access_token, refresh_token, expires_in, token_type, user]
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
          example: "Bearer"
      required: [access_token, expires_in, token_type]

    # Report schemas
    WorkSummaryReport:
      type: object
      properties:
        summary:
          type: object
          properties:
            total_work_logs:
              type: integer
            total_cost:
              type: number
              format: float
            average_cost:
              type: number
              format: float
            completion_rate:
              type: number
              format: float
              description: Percentage of completed work logs
        data:
          type: array
          items:
            type: object
            properties:
              group_key:
                type: string
                description: The grouping field value
              work_count:
                type: integer
              total_cost:
                type: number
                format: float
              average_cost:
                type: number
                format: float
              completed_count:
                type: integer
              critical_count:
                type: integer
        filters:
          type: object
          description: Applied filters for the report
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Number of items per page
        has_more:
          type: boolean
          description: Whether there are more items available
        next_cursor:
          type: string
          nullable: true
          description: Cursor for the next page
      required: [total, limit, has_more]
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          nullable: true
          description: Additional error details
        request_id:
          type: string
          description: Unique request identifier for debugging
      required: [error, message]
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "One or more fields are invalid"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field name with validation error
              code:
                type: string
                description: Validation error code
              message:
                type: string
                description: Human-readable validation error message
            required: [field, code, message]
        request_id:
          type: string
      required: [error, message, details]
...